+++ cmd/flags.go
@@ -4,9 +4,11 @@ import (
 	"fmt"
 	"log/slog"
 	"os"
+	"strings"
 
 	bhlquest "github.com/gnames/bhlquest/pkg"
 	"github.com/gnames/bhlquest/pkg/config"
+	"github.com/gnames/gnlib"
 	"github.com/spf13/cobra"
 )
 
@@ -19,6 +21,13 @@ func debugFlag(cmd *cobra.Command) {
 		handle := slog.NewJSONHandler(os.Stderr, opts)
 		logger := slog.New(handle)
 		slog.SetDefault(logger)
+	}
+}
+
+func noConfirmFlag(cmd *cobra.Command) {
+	b, _ := cmd.Flags().GetBool("yes-to-confirmations")
+	if b {
+		opts = append(opts, config.OptWithoutConfirm(true))
 	}
 }
 
@@ -39,5 +48,16 @@ func versionFlag(cmd *cobra.Command) {
 			version.Build,
 		)
 		os.Exit(0)
+	}
+}
+
+func taxaFlag(cmd *cobra.Command) {
+	s, _ := cmd.Flags().GetString("taxa")
+	if s != "" {
+		el := strings.Split(s, ",")
+		taxa := gnlib.Map(el, func(s string) string {
+			return strings.TrimSpace(s)
+		})
+		opts = append(opts, config.OptInitTaxa(taxa))
 	}
 }